controller is ready

+++ server/services/rentaCards.service.js
@@ -20,14 +20,32 @@ class RentaService {
             throw new Error(error.message) 
         }
     }
-    static async deleteRentaCard(id,){
+    static async deleteRentaCard(id, user_id){
         try {
-            const deleteRentaCard = await Renta.destroy({where: {id}})
+            const deleteRentaCard = await Renta.destroy({where: {id, user_id: 1}})
             return deleteRentaCard;
         } catch (error) {
             throw new Error(error.message);
         }
     }
-
+    static async updateRenta(data,id){
+        try {
+            const [countUpdated] = await Renta.update(data, { where: {id} })
+            return countUpdated
+        } catch (error) {
+            throw new Error(error.message);
+        }
+    }
+    static async findRentaByName(name){
+        try {
+            const renta = await Renta.findByPk({where: { name }})
+            return renta
+        } catch (error) {
+            throw new Error(error.message);
+        }
+    }
     
 }
+
+
+module.exports = RentaService;
